Major refactor, new system
Hit the V key to teleport to new, experimental randomized system.  Not much to do there yet, but it seems to have working planets and moons.  The targeting system / ship radar has been disabled because it's jank as fuck and is causing problems in the new system (more so than in the home system, where it always caused problems... sometimes).