Job status is accurate
HUD now shows "current" job (last taken job), as well as how many jobs the player has taken in the system, changing to a different job when the current job is complete.  Still no way to select which job is listed as "current" or show all jobs taken in the system.